feature flagged blitz-shell behind new shell feature
