+++ packages/native/src/shell/assets.rs
