+++ packages/native/src/shell/link_handler.rs
